requirements.txt
# Core
openai==2.7.1
langchain==1.0.4
langgraph==1.0.2
# langchain-core==1.*
langchain-openai==1.0.2
langchain-community==0.4.1

# Vector store / RAG
faiss-cpu==1.8.0.post1
pydantic==2.*

# Optional utils
python-dotenv==1.0.1
rich==13.*
